# Add Generics Support, More Idiomatic Get()