@adrianosela 2025-02-04
Any thoughts on this @dghubble? Wondering if there is interest in this change or if I should close the PR. Feel free to close it if not.